Added view module in elixir
